third_party/WebKit/LayoutTests/fast/forms/input-stepup-stepdown.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('Check stepUp() and stepDown() bahevior for unsupported types.');

var input = document.createElement('input');

debug('Unsupported type');
input.type = 'text';
shouldThrow('input.step = "3"; input.min = ""; input.max = ""; input.value = "2"; input.stepDown()', '"InvalidStateError: Failed to execute \'stepDown\' on \'HTMLInputElement\': This form element is not steppable."');
shouldThrow('input.stepDown(0)', '"InvalidStateError: Failed to execute \'stepDown\' on \'HTMLInputElement\': This form element is not steppable."');
shouldThrow('input.stepUp()', '"InvalidStateError: Failed to execute \'stepUp\' on \'HTMLInputElement\': This form element is not steppable."');
shouldThrow('input.stepUp(0)', '"InvalidStateError: Failed to execute \'stepUp\' on \'HTMLInputElement\': This form element is not steppable."');
</script>
</body>
</html>
